refactor(backend): validate the lock service connection without cast()
Review feedback: `cast()` is not an acceptable narrowing tool — it is an
unchecked assertion in the same family as an inline `type: ignore`. Replace it
with the real check.

`_require_services_connection()` now positively checks
`isinstance(connection, InfrahubServices)`, so an unsupported object is rejected
at construction with the driver named instead of reaching `NATSLock` and failing
as an `AttributeError` on first acquire. No `cast`, and the docstring no longer
has to explain one.

That check needs the class at runtime, which closes the `infrahub.services` ->
`infrahub.lock` import cycle and made mypy bind the `infrahub.core.registry`
submodule instead of the re-exported singleton (15 errors in `infrahub/types.py`
and `infrahub/tasks/registry.py`). Fixed at the root: both modules now import
the singleton directly, as `tasks/registry.py` already did for
`infrahub.graphql.registry`. Same object at runtime, no longer order-dependent.

Extends the guard test to cover an arbitrary unsupported object, which is the
case that distinguishes a real check from narrowing by elimination.

Ref: INBOX-32
